requirements.txt
fastapi
uvicorn
sqlalchemy
aiohttp
beautifulsoup4
playwright
python-dotenv
httpx
